+++ GerenciamentoInvestimentos.Domain/Interfaces/Repositories/IUserRepository.cs
@@ -6,6 +6,6 @@ public interface IUserRepository
 {
     long Create(User user);
     IEnumerable<User> Get();
-    User Get(int id);
+    User Get(long id);
     User GetByEmail(string email);
 }
